maru: Enable in-tree kernel builds
Note that since LOCAL_KERNEL is defined in device.mk, we do not have
access to variables like $(TARGET_OUT_INTERMEDIATES) so we are forced to
hardcode in the out directory.

This will also not work with custom build output directory, but as far
as I know no one uses that, and the AndroidKernel.mk in
android_kernel_hammerhead will also need to be updated to support custom
output directories.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>