maru: Add maru_hammerhead lunch combos
We add the lunch combos in the device tree now instead of vendor/maruos.

The aosp_hammerhead combos are removed because they don't build with the
Maru OS tree due to restricted vendor files. This simplifies the lunch
combo options a bit; ideally, the only lunch options listed should be
valid products for Maru OS.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>