maru: Simplify maru_hammerhead product config
This inherits some stuff required for LineageOS like the
BUILD_FINGERPRINT override.

I learnt the hard way that the BUILD_FINGERPRINT is necessary to build a
signed update zip that actually boots the framework...without it there's
an error about no ro.build.fingerprint set and a subsequent fatal error
that causes the framework to restart in a loop.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>